chore: relocate SlotSummary to cproto
